% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_setup.R
\name{arg_combinations}
\alias{arg_combinations}
\title{Get all combinations of each named argument}
\usage{
arg_combinations(...)
}
\arguments{
\item{...}{arguments to get combinations of; splicable with !!!}
}
\value{
a tibble of all argument combinations
}
\description{
Pass a formula or list of formulas to delay evaluation until after
after combinations have been made.
}
\examples{
# Use formulas to create a function of other combinations
arg_combinations(
  left = 1:3,
  right = 1:3,
  left_greater = ~ left > right)

# Use lists...


}
