% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_sbatch}
\alias{assert_sbatch}
\title{Assert a sbtach object}
\usage{
assert_sbatch(x)
}
\arguments{
\item{x}{a object to assert is a \code{sbatch} object.}
}
\description{
Assert a sbtach object
}
