% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{cat_status}
\alias{cat_status}
\title{Prints a status while running an expression}
\usage{
cat_status(.code, .status)
}
\arguments{
\item{.code}{an expression to run}

\item{.status}{the status to print}
}
\value{
the value of \code{.code}
}
\description{
Prints a status while running an expression
}
