% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOB_STATE_CODES.R
\docType{data}
\name{JOB_STATE_CODES}
\alias{JOB_STATE_CODES}
\title{Slurm Job state codes}
\format{A data frame with 24 rows and 4 columns.}
\usage{
JOB_STATE_CODES
}
\description{
This data frame contains information regarding the job state codes that Slurm
returns when querying the status of a given job. The last column, \code{type},
shows a description of how that corresponding state is considered in the
package's various operations. This is used in the function \link{status}.
}
\references{
Slurm's website \url{https://slurm.schedmd.com/squeue.html}
}
\keyword{datasets}
