% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_files.R
\name{make_slurm_config}
\alias{make_slurm_config}
\title{Create a SLURM configuration header}
\usage{
make_slurm_config(job_name, nodes, wayness, time, queue, account, log_dir,
  email = "", ...)
}
\arguments{
\item{job_name}{the job's name}

\item{nodes}{number of nodes}

\item{wayness}{how many tasks per node to use}

\item{time}{max clock time in "HH:MM:SS" format}

\item{queue}{(current options are "normal" and "development")}

\item{account}{account to charge the jobs to}

\item{log_dir}{where to save the log files}

\item{email}{address to email}

\item{...}{unused}
}
\value{
the config header of a slurm job
}
\description{
Create a SLURM configuration header
}
