% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_swarm.R
\name{make_slurm_swarm}
\alias{make_slurm_swarm}
\title{Factory for making a slurm_swarm}
\usage{
make_slurm_swarm(
  swarm_name,
  updater_script,
  remaining_jobs_script,
  section_list,
  ppn_list,
  queue_list = c("normal", "skx-normal"),
  jobid = "$SLURM_JOBID",
  other_args = list(),
  sbatch_extra = ""
)
}
\arguments{
\item{swarm_name}{name of the swarm.}

\item{updater_script}{name of the command script used to update task lists.
This script should take \code{section_list} as the first
argument and the SLURM job ID as the second.}

\item{remaining_jobs_script}{name of command script used to determine how many
jobs are remaining.}

\item{section_list}{vector of job types.}

\item{ppn_list}{processes per node for \code{section_list}.}

\item{queue_list}{available slurm partitions (queues)}

\item{jobid}{environment variable for the slurm job ID}

\item{other_args}{named list of other args w/ default values for; this can
used to set default values for any existing arguments other
than section or queue. Any new args included here will only
be referenced in \code{sbatch_extra}, \code{updater_script},
\code{remaining_jobs_script}, or \code{update_arg_list}.}

\item{sbatch_extra}{extra #SBATCH directives}

\item{update_arg_list}{character vector of a section's extra \code{glue}-able
args for the \code{updater_script}.}

\item{nodes}{number of nodes}

\item{time}{time requested}

\item{queue}{which queue to use}

\item{sections}{which job types to run}

\item{account}{TACC account}

\item{email}{email for contact}

\item{root_dir}{directory to create swarm folder in}
}
\value{
a function that sets up a slurm swarm
}
\description{
Factory for making a slurm_swarm
}
