% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_update.R
\name{make_swarm_task_helpers}
\alias{make_swarm_task_helpers}
\title{Swarm updater script}
\usage{
make_swarm_task_helpers(
  swarm_name,
  section_list,
  script_list,
  filter_function_list,
  script_arg_list,
  in_dir_list,
  in_pat_list,
  out_dir_list,
  out_pat_list
)
}
\arguments{
\item{swarm_name}{name of the swarm script}

\item{section_list}{vector of section names}

\item{script_list}{vector of scripts for each name}

\item{filter_function_list}{list of functions or lambdas that filter out completed
tasks; arguments should be .x for input files and
.y for output files. return value should be the
first arguments for the task's script}

\item{script_arg_list}{extra arguments for the scripts}

\item{in_dir_list, out_dir_list}{directories to scan for input and output files}

\item{in_pat_list, out_pat_list}{regex pattern to filter input/output file names by}
}
\value{
a list with the updater function and remaining function needed by the swarm
}
\description{
Swarm updater script
}
