% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSlurmCluster.R
\name{makeSlurmCluster}
\alias{makeSlurmCluster}
\alias{stopCluster.SlurmCluster}
\title{Create a Parallel Socket Cluster using Slurm}
\usage{
makeSlurmCluster(njobs, job_name = opts_sluRm$get_job_name(),
  tmp_path = opts_sluRm$get_tmp_path(), sbatch_opt = list(),
  timeout = 300L, ...)

\method{stopCluster}{SlurmCluster}(cl)
}
\arguments{
\item{njobs}{Integer. Number of jobs to use in the job-array.}

\item{job_name}{Character. Name of the job to be passed to \code{Slurm}.}

\item{tmp_path}{Character. Path to the directory where all the data (including
scripts) will be stored. Notice that this path must be accessible by all the
nodes in the network (See \link{opts_sluRm}).}
}
\description{
Create a Parallel Socket Cluster using Slurm
}
\details{
By default, if the \code{time} option is not specified via \code{sbatch_opt},
then it is set to the value \code{1-0}, this is, 1 day and 0 hours.
}
