% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{new_slurm_job}
\alias{new_slurm_job}
\alias{slurm_job}
\title{Slurm jobs}
\usage{
new_slurm_job(call, rscript, bashfile, robjects, njobs, job_opts)
}
\arguments{
\item{call}{The original call}

\item{rscript, bashfile}{The R script and bash file path.}

\item{robjects}{A character vector of R objects that will be imported in the job.}

\item{njobs}{Integer. Number of jobs to start (array).}

\item{job_opts}{List. Arguments to be passed to \link{sbatch}.}
}
\description{
Slurm jobs
}
