% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.print.R
\name{num_to_time}
\alias{num_to_time}
\title{Convert hours in numerical format to slurm time format}
\usage{
num_to_time(x)
}
\arguments{
\item{x}{Hours as numeric.}
}
\value{
Hours in slurm format
}
\description{
Convert hours in numerical format to slurm time format
}
