% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{options_sluRm}
\alias{options_sluRm}
\title{Get and set default options for slurm}
\format{An object of class \code{environment} of length 2.}
\usage{
options_sluRm
}
\description{
Get and set default options for slurm
}
\details{
Current supported options are:
\itemize{
\item \code{job_path}: This can be retrieved as\preformatted{options_sluRm$get_job_path()
}

Likewise it can be set as\preformatted{options_sluRm$get_job_path("some_path")
}

Currently, the default value is set when loading the package and is the
current working directory.
\item ``
}
}
\keyword{datasets}
