% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{opts_sluRm}
\alias{opts_sluRm}
\title{Get and set default options for slurm}
\format{An object of class \code{opts_sluRm} of length 10.}
\usage{
opts_sluRm
}
\description{
Get and set default options for slurm
}
\details{
Current supported options are:
\itemize{
\item \code{chdir}: This can be retrieved as\preformatted{opts_sluRm$get_chdir()
}

Likewise it can be set as\preformatted{opts_sluRm$get_chdir("some_path")
}

Currently, the default value is set when loading the package and is the
current working directory.
\item ``
}
}
\keyword{datasets}
