% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_flags}
\alias{parse_flags}
\alias{parse_flags.default}
\alias{parse_flags.list}
\title{Utility function}
\usage{
parse_flags(...)

\method{parse_flags}{default}(...)

\method{parse_flags}{list}(x, ...)
}
\arguments{
\item{...}{Options to be parsed as bash flags.}

\item{x}{A named list.}
}
\value{
A character vector with the processed flags.
}
\description{
Utility function
}
\examples{
cat(parse_flags(a=1, b=TRUE, hola=2, y="I have spaces", ms=2, `cpus-per-task`=4))
# -a 1 -b --hola=2 -y "I have spaces" --ms=2 --cpus-per-task=4
}
\seealso{
Other utilities: 
\code{\link{Slurm_clean}()},
\code{\link{Slurm_env}()},
\code{\link{Slurm_log}()},
\code{\link{WhoAmI}()},
\code{\link{snames}()},
\code{\link{status}()}
}
\concept{utilities}
