% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{require_globals}
\alias{require_globals}
\title{Checks to make sure that global variables are defined}
\usage{
require_globals(...)
}
\arguments{
\item{...}{unquoted global parameters}
}
\value{
invisible
}
\description{
Checks to make sure that global variables are defined
}
