% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{require_names}
\alias{require_names}
\title{Ensure that global variables have (only) the right names}
\usage{
require_names(names, ...)
}
\arguments{
\item{names}{required names}

\item{...}{global variables to check}
}
\description{
Ensure that global variables have (only) the right names
}
