% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rscript_header}
\alias{rscript_header}
\title{Creates an R script}
\usage{
rscript_header(pkgs = loaded_packages())
}
\arguments{
\item{pkgs}{A named list of R packages to load.}
}
\description{
Creates an R script
}
