% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.R
\name{sbatch}
\alias{sbatch}
\alias{sbatch_seq}
\title{submit a slurm file to the queue, with dependencies}
\usage{
sbatch(
  slurm_file,
  dep = NA,
  dep_type = c("ok", "notok", "any", ""),
  .extra = ""
)

sbatch_seq(slurm_files, ...)
}
\arguments{
\item{slurm_file}{name of the file to submit}

\item{dep_type}{which type of dependency to use (see SLURM documentation)}

\item{slurm_files}{a vector of slurm files to be submitted in order}

\item{...}{arguments for \code{sbatch}}

\item{dependencies}{jobid of dependencies if any}

\item{extra}{additinoal SLURM arguments}
}
\value{
the jobid of the submitted job

a vector of jobids
}
\description{
submit a slurm file to the queue, with dependencies

submit a sequence of slurm files, each depending on the previous
}
