% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sluRm.R
\docType{package}
\name{sluRm}
\alias{sluRm}
\alias{sluRm-package}
\title{A lightweight wrapper for HPC with Slurm
A lightweight wrapper for HPC with Slurm}
\description{
A lightweight wrapper for HPC with Slurm
A lightweight wrapper for HPC with Slurm
}
