% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{slurm_available}
\alias{slurm_available}
\title{Check whether Slurm is available in the system}
\usage{
slurm_available()
}
\value{
A Logical scalar indicating whether Slurm is available (\code{TRUE}) or not
(\code{FALSE})
}
\description{
Check whether Slurm is available in the system
}
\examples{
slurm_available()

}
