% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Slurm_clean}
\alias{Slurm_clean}
\title{Clean a session.}
\usage{
Slurm_clean(x)
}
\arguments{
\item{x}{An object of class \code{slurm_job}.}
}
\description{
Clean a session.
}
\seealso{
Other post submission: \code{\link{Slurm_collect}},
  \code{\link{Slurm_log}}, \code{\link{state}}

Other utilities: \code{\link{Slurm_env}},
  \code{\link{Slurm_log}}, \code{\link{WhoAmI}},
  \code{\link{parse_flags}}, \code{\link{snames}},
  \code{\link{state}}
}
\concept{post submission}
\concept{utilities}
