% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{Slurm_collect}
\alias{Slurm_collect}
\alias{Slurm_collect.slurm_call}
\title{Collect the results}
\usage{
Slurm_collect(...)

\method{Slurm_collect}{slurm_call}(x, ...)
}
\arguments{
\item{...}{Further arguments passed to the method.}

\item{x}{An object of class \code{slurm_call}.}
}
\description{
Collect the results
}
