% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Slurm_env}
\alias{Slurm_env}
\title{A wrapper of \link{Sys.getenv}}
\usage{
Slurm_env(x)
}
\arguments{
\item{x}{Character scalar. Environment variable to get.}
}
\description{
A wrapper of \link{Sys.getenv}
}
