% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\name{Slurm_lapply}
\alias{Slurm_lapply}
\title{The Slurm version of the \code{lapply}  function}
\usage{
Slurm_lapply(X, fun, ..., job_name = getOption("sluRm.job_name", "sluRm"),
  path = tempdir())
}
\arguments{
\item{X, fun, ...}{See \link{lapply}.}

\item{job_name}{Character. Name of the job to be passed to `Slurm`.}

\item{job_path}{Character. Path to the directory where all the data (including
scripts) will be stored.}
}
\description{
The Slurm version of the \code{lapply}  function
}
