% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slurm_log.R
\name{Slurm_log}
\alias{Slurm_log}
\title{Check the R logfile of a job.}
\usage{
Slurm_log(x, cmd = "less", which. = NULL)
}
\arguments{
\item{x}{An object of class \link{slurm_job}.}

\item{cmd}{Character scalar. The name of the command to use to call view the
log file.}

\item{which.}{An integer scalar. The number of the array job to check. This
should range between 1 and \code{x$njobs}.}
}
\description{
Check the R logfile of a job.
}
\examples{
\dontrun{
x <- Slurm_EvalQ(sluRm::whoami(), plan = "wait")
Slurm_log(x) # Checking the R log
}
}
\seealso{
Other post submission: \code{\link{Slurm_clean}},
  \code{\link{Slurm_collect}}, \code{\link{state}}

Other utilities: \code{\link{Slurm_clean}},
  \code{\link{Slurm_env}}, \code{\link{WhoAmI}},
  \code{\link{parse_flags}}, \code{\link{snames}},
  \code{\link{state}}
}
\concept{post submission}
\concept{utilities}
