% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmr-package.R
\docType{package}
\name{slurmr-package}
\alias{slurmr}
\alias{slurmr-package}
\title{slurmr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
