% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmR.R
\docType{package}
\name{slurmR}
\alias{slurmR}
\title{A Lightweight Wrapper for 'Slurm'}
\description{
'Slurm', Simple Linux Utility for Resource Management
\url{https://slurm.schedmd.com/}, is a popular 'Linux' based software used to
schedule jobs in 'HPC' (High Performance Computing) clusters. This R package
provides a specialized lightweight wrapper of 'Slurm' with a syntax similar to
that found in the 'parallel' R package. The package also includes a method for
creating socket cluster objects spanning multiple nodes that can be used with
the 'parallel' package.
}
\details{
To cite slurmR in publications use:

Vega Yon et al., (2019). slurmR: A lightweight wrapper for HPC with Slurm. Journal of Open Source Software,
4(39), 1493, https://doi.org/10.21105/joss.01493

A BibTeX entry for LaTeX users is\preformatted{ @Article\{,
  title = \{slurmR: A lightweight wrapper for HPC with Slurm\},
  author = \{George \{Vega Yon\} and Paul Marjoram\},
  journal = \{The Journal of Open Source Software\},
  year = \{2019\},
  month = \{jul\},
  volume = \{4\},
  number = \{39\},
  doi = \{10.21105/joss.01493\},
  url = \{https://doi.org/10.21105/joss.01493\},
\}
}
}
