% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sluRm.R
\docType{package}
\name{slurrm}
\alias{slurrm}
\alias{slurrm-package}
\title{A Lightweight Wrapper for 'Slurm'}
\description{
A Lightweight Wrapper for 'Slurm'
}
\details{
To cite slurrm in publications use:

Vega Yon et al., (2019). sluRm: A lightweight wrapper for HPC with Slurm. Journal of Open Source Software,
4(39), 1493, https://doi.org/10.21105/joss.01493

A BibTeX entry for LaTeX users is\preformatted{ @Article{,
  title = {sluRm: A lightweight wrapper for HPC with Slurm},
  author = {George {Vega Yon} and Paul Marjoram},
  journal = {The Journal of Open Source Software},
  year = {2019},
  month = {jul},
  volume = {4},
  number = {39},
  doi = {10.21105/joss.01493},
  url = {https://doi.org/10.21105/joss.01493},
}
}
}
