% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\name{squeue}
\alias{squeue}
\title{Return SLURM queue as a tibble}
\usage{
squeue(user = "$USER")
}
\arguments{
\item{user}{username}
}
\value{
a tibble of the SLURM queue
}
\description{
Return SLURM queue as a tibble
}
