% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{state}
\alias{state}
\alias{state.slurm_job}
\alias{JOB_STATE_CODES}
\alias{state.default}
\title{Check the State (status) of a Slurm JOB}
\format{An object of class \code{data.frame} with 24 rows and 4 columns.}
\usage{
state(x)

\method{state}{slurm_job}(x)

JOB_STATE_CODES

\method{state}{default}(x)
}
\arguments{
\item{x}{Either a Job id, or an object of class \code{slurm_job}.}
}
\value{
An integer with attributes. The attributes are integer vectors indicating
which jobs fail in the categories of \code{completed}, \code{failed}, and \code{pending} (see
\link{JOB_STATE_CODES}). Possible return values are
\itemize{
\item \code{-1}: Job not found. This may be a false negative since Slurm may still
be scheduling the job.
\item \code{0}: Job completed. In this case all the components of the job
have finalized without any errors.
\item \code{1}: Some jobs are still running/pending.
\item \code{2}: One or more jobs have failed.
}

If the job is not an array, then function will return the corresponding code
but the attributes will only have a single number, 1, according to the state
of the job (completed, failed, pending).
}
\description{
Using the \link{sacct} function, it checks the status of a particular error and
returns information about its current state.
}
\examples{
\dontrun{

x <- Slurm_EvalQ(Sys.sleep(100), njobs = 2)

state(x)
# A possible result: An integer with attributes
# [1] 1
# attr(,"pending")
# [1] 1
# attr(,"failed")
# integer(0)
# attr(,"done")
# [1] 2


}
}
\seealso{
Other utilities: \code{\link{Slurm_clean}},
  \code{\link{Slurm_env}}, \code{\link{Slurm_log}},
  \code{\link{WhoAmI}}, \code{\link{parse_flags}},
  \code{\link{snames}}

Other post submission: \code{\link{Slurm_clean}},
  \code{\link{Slurm_collect}}, \code{\link{Slurm_log}}
}
\concept{post submission}
\concept{utilities}
\keyword{datasets}
