% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{unnest_args}
\alias{unnest_args}
\title{unnests arguments to be saved in a database
This is helpful for environmental & popsize args that hav multiple populations}
\usage{
unnest_args(.data)
}
\arguments{
\item{.data}{a tibble of arguments}
}
\value{
an unnested data frame, with an extra sub_pop column
}
\description{
unnests arguments to be saved in a database
This is helpful for environmental & popsize args that hav multiple populations
}
