% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WhoAmI}
\alias{WhoAmI}
\title{Information about where jobs are submitted}
\usage{
WhoAmI()
}
\description{
This returns a named vector with the following variables:
\Sexpr{paste(names(sluRm::WhoAmI()), collapse = ", ")}
}
