% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{new_slurm_job}
\alias{new_slurm_job}
\title{Slurm Jobs}
\usage{
new_slurm_job(call, rscript, bashfile, robjects, job_name, job_path, njobs)
}
\arguments{
\item{call}{The original call}

\item{rscript, bashfile}{The R script and bash file path.}

\item{robjects}{A character vector of R objects that will be imported in the job.}

\item{job_name, job_path}{Character. Name and path of the job.}

\item{njobs}{Integer. Number of jobs to start (array).}
}
\description{
Slurm Jobs
}
