% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{op-na-default}
\alias{op-na-default}
\alias{\%|\%}
\title{Replace missing values}
\usage{
x \%|\% value_if_x_is_na
}
\description{
See \verb{rlang::[\\\%|\\\%][rlang::op-na-default]} for details.
}
\keyword{internal}
