% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{opts_sluRm}
\alias{opts_sluRm}
\title{Get and set default options for \code{sbatch} and \code{sluRm} internals}
\format{An object of class \code{opts_sluRm} of length 13.}
\usage{
opts_sluRm
}
\description{
Most of the functions in the \code{sluRm} package use \code{chdir} and \code{job-name}
options to write and submit jobs to \strong{Slurm}. These options have global
defaults that are set and retrieved using \code{opts_sluRm}.
}
\details{
Current supported options are:

Debuggin mode
\itemize{
\item \code{debug_on : function ()} \Sexpr{attr(sluRm::opts_sluRm$debug_on, "desc")}
\item \code{debug_off : function ()} \Sexpr{attr(sluRm::opts_sluRm$debug_off, "desc")}
\item \code{get_debug : function ()} \Sexpr{attr(sluRm::opts_sluRm$get_debug, "desc")}
}

Verbose mode
\itemize{
\item \code{verbose_on : function ()} \Sexpr{attr(sluRm::opts_sluRm$verbose_on, "desc")}
\item \code{verbose_off : function ()} \Sexpr{attr(sluRm::opts_sluRm$verbose_off, "desc")}
\item \code{get_verbose : function ()} \Sexpr{attr(sluRm::opts_sluRm$get_verbose, "desc")}
}

Slurm options
\itemize{
\item \code{set_chdir : function (path, recursive = TRUE)} \Sexpr{attr(sluRm::opts_sluRm$set_chdir, "desc")}
\item \code{get_chdir : function ()} \Sexpr{attr(sluRm::opts_sluRm$get_chdir, "desc")}
\item \code{set_job_name : function (path, check = TRUE, overwrite = TRUE)} \Sexpr{attr(sluRm::opts_sluRm$set_job_name, "desc")}.
\item \code{get_job_name : function (check = TRUE)} \Sexpr{attr(sluRm::opts_sluRm$get_job_name, "desc")}
}

Other options
\itemize{
\item \code{get_cmd : function ()} \Sexpr{attr(sluRm::opts_sluRm$get_cmd, "desc")}
}

For general set/retrieve options
\itemize{
\item \code{set_opts : function (...)} \Sexpr{attr(sluRm::opts_sluRm$set_opts, "desc")}
\item \code{get_opts : function (...)} \Sexpr{attr(sluRm::opts_sluRm$get_opts, "desc")}
}
}
\examples{

# Common setup
\dontrun{
opts_sluRm$set_chdir("/staging/pdt/vegayon")
opts_sluRm$set_job_name("simulations-1")
opts_slurm$set_opts(partition="thomas", account="lc_pdt")
}

}
\keyword{datasets}
