% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.R
\name{sbatch}
\alias{sbatch}
\title{Sbatch object constructor}
\usage{
sbatch(code = "", add_status = TRUE, path = "", J_ = NULL, t_ = 1,
  A_ = NULL, n_ = NULL, mem__ = NULL, mail_type__ = NULL,
  mail_user__ = NULL, output__ = "slurm-\%j")
}
\arguments{
\item{code}{Code to run in SBATCH. Every row in batch script is a char element.}

\item{add_status}{Add code to print out status for run.}

\item{path}{Path where write out SBATCH files e.t.c.}

\item{_J}{See details.}

\item{_t}{See details.}

\item{_A}{See details.}

\item{_n}{See details.}

\item{_n}{See details.}

\item{__mem}{See details.}

\item{__mail_type}{See details.}

\item{__output}{See details.}
}
\value{
\code{sbatch} object
}
\description{
Sbatch object constructor
}
\details{
See slurm details om arguments. Use \code{sbatch -help} for details. 
The parameters implemented can be seen below.
}

