% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{sbatch}
\alias{sbatch}
\alias{sbatch.slurm_job}
\alias{squeue}
\alias{squeue.slurm_job}
\alias{scancel}
\alias{scancel.slurm_job}
\alias{sacct}
\alias{sacct.slurm_job}
\alias{sacct.default}
\title{A wrapper of \code{sbatch}.}
\usage{
sbatch(x, wait = TRUE, submit = TRUE, ...)

\method{sbatch}{slurm_job}(x, wait = TRUE, submit = TRUE, ...)

squeue(x, ...)

\method{squeue}{slurm_job}(x, ...)

scancel(x, ...)

\method{scancel}{slurm_job}(x, ...)

sacct(x, ...)

\method{sacct}{slurm_job}(x, ...)

\method{sacct}{default}(x, brief = TRUE, parsable = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{slurm_job}.}

\item{wait}{Logical scalar. When \code{TRUE} the function will pass the \code{--wait}
flag to \code{Slurm} and set \code{wait=TRUE} in the \link{system2} call.}

\item{submit}{Logical, when \code{TRUE} calls \link{sbatch} to submit the job to slurm.}

\item{...}{Further flags passed to the command line function.}

\item{brief}{Logical. Passed to \code{sacct}.}

\item{parsable}{Logical. Passed to \code{sacct}.}
}
\description{
A wrapper of \code{sbatch}.
}
