% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{sbatch}
\alias{sbatch}
\alias{sbatch.slurm_job}
\alias{squeue}
\alias{squeue.slurm_job}
\alias{scancel}
\alias{scancel.slurm_job}
\title{A wrapper of \code{sbatch}.}
\usage{
sbatch(x, wait = TRUE, ...)

\method{sbatch}{slurm_job}(x, wait = TRUE, ...)

squeue(x, ...)

\method{squeue}{slurm_job}(x, ...)

scancel(x, ...)

\method{scancel}{slurm_job}(x, ...)
}
\arguments{
\item{x}{An object of class \code{slurm_job}.}

\item{wait}{Logical scalar. When \code{TRUE} the function will pass the \code{--wait}
flag to \code{Slurm} and set \code{wait=TRUE} in the \link{system2} call.}

\item{...}{Further flags passed to the command line function.}
}
\description{
A wrapper of \code{sbatch}.
}
