% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.R
\name{sbatch}
\alias{sbatch}
\alias{sbatch.slurm_job}
\alias{scancel}
\alias{scancel.slurm_job}
\title{A wrapper of \code{sbatch}.}
\usage{
sbatch(x)

\method{sbatch}{slurm_job}(x)

scancel(x)

\method{scancel}{slurm_job}(x)
}
\arguments{
\item{x}{An object of class \code{slurm_job}.}
}
\description{
A wrapper of \code{sbatch}.
}
