% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{Slurm_collect}
\alias{Slurm_collect}
\alias{Slurm_collect.slurm_job}
\title{Collect the results}
\usage{
Slurm_collect(...)

\method{Slurm_collect}{slurm_job}(x, any = TRUE, ...)
}
\arguments{
\item{...}{Further arguments passed to the method.}

\item{x}{An object of class \code{slurm_call}.}

\item{any}{Logical. When \code{TRUE}, will collect any output available regardless
of whether the job is completed or not.}
}
\description{
Collect the results
}
\examples{
\dontrun{
# Collecting a job after calling it
job <- Slurm_EvalQ(sluRm::WhoAmI(), njobs = 4, plan = "wait")
Slurm_collect(job)

# Collecting a job from a previous R session
job <- read_slurm_job("/path/to/a/job/tmp_dir")
Slurm_collect(job)
}
}
