% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{Slurm_collect}
\alias{Slurm_collect}
\alias{Slurm_collect.slurm_call}
\title{Collect the results}
\usage{
Slurm_collect(...)

\method{Slurm_collect}{slurm_call}(x, ...)
}
\description{
Collect the results
}
