% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalq.R
\name{Slurm_EvalQ}
\alias{Slurm_EvalQ}
\title{Submit an expression to be evaluated to multiple jobs.}
\usage{
Slurm_EvalQ(expr, njobs = 2L, job_name = opts_sluRm$get_job_name(),
  job_path = opts_sluRm$get_chdir(), plan = "collect",
  sbatch_opt = list(ntasks = 1L), rscript_opt = list(vanilla = TRUE),
  seeds = 1L:njobs, compress = TRUE, export = NULL,
  libPaths = .libPaths(), hooks = NULL)
}
\arguments{
\item{expr}{An expression to be passed to Slurm.}

\item{njobs}{Integer. Number of jobs to specity.}

\item{job_name}{Character. Name of the job to be passed to \code{Slurm}.}

\item{job_path}{Character. Path to the directory where all the data (including
scripts) will be stored (See \link{opts_sluRm}).}

\item{plan}{A character scalar. (See \link{the_plan}).}

\item{sbatch_opt, rscript_opt}{List. Options to be passed via flags to
the bash file as \code{#SBATCH} and to \code{Rscript} respectively.}

\item{seeds}{Integer vector of length \code{njobs}. Seeds to be passed to each
job.}

\item{compress}{Logical scalar (default \code{TRUE}). Passed to \link{saveRDS}. Setting
this value to \code{FALSE} can be useful when the user requires faster read/write
of R objects on disk.}

\item{export}{A named list with objects to be included in the Spawned sessions.}

\item{libPaths}{A character vector. See \link{.libPaths}.}

\item{hooks}{A list of functions (passed to \link{new_slurm_job}).}
}
\description{
Submit an expression to be evaluated to multiple jobs.
}
