% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snames}
\alias{snames}
\title{Full path names for Slurm jobs}
\usage{
snames(type, array_id)
}
\arguments{
\item{type}{can be any of r, sh, out, or rds.}

\item{array_id}{Integer. ID of the array to create the name.}
}
\description{
Using \link{opts_sluRm}\code{$get_chdir} and \link{opts_sluRm}\code{$get_job_name} creates
file names with full path to the objects. This function is intended for
internal use only.
}
\seealso{
Other Utility: \code{\link{parse_flags}},
  \code{\link{state}}
}
\concept{Utility}
