% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSlurm.R
\name{sourceSlurm}
\alias{sourceSlurm}
\title{Source an R script as a Slurm job}
\usage{
sourceSlurm(file, job_name = NULL, tmp_path = tempdir(),
  rscript_opt = list(vanilla = TRUE), sbatch_opt = list(), ...)
}
\arguments{
\item{file}{Character. Path to the R script to source using Slurm.}

\item{job_name}{Character. Name of the job to be passed to \code{Slurm}.}

\item{tmp_path}{Character. Path to the directory where all the data (including
scripts) will be stored. Notice that this path must be accessible by all the
nodes in the network (See \link{opts_sluRm}).}

\item{rscript_opt}{List. Options to be passed to \code{Rscript}.}

\item{sbatch_opt}{List of options to be passed to \code{sbatch}. This is usually
done by adding the flags \code{#SBATCH} to the bash file.}

\item{...}{Further arguments to be passed to \link{sbatch}.}
}
\description{
This function sources R scripts using Slurm by creating a batch script file
and submitting it via \link{sbatch}.
}
