% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{state}
\alias{state}
\alias{state.slurm_job}
\alias{JOB_STATE_CODES}
\alias{state.default}
\title{Check the State (status) of a Slurm JOB}
\format{An object of class \code{data.frame} with 24 rows and 3 columns.}
\usage{
state(x)

\method{state}{slurm_job}(x)

JOB_STATE_CODES

\method{state}{default}(x)
}
\arguments{
\item{x}{Either a Job id, or an object of class \code{slurm_job}.}
}
\value{
An integer with attributes. The attributes are integer vectors indicating
which jobs fail in the categories of \code{done}, \code{failed}, \code{running}, and \code{pending}.
The value of the integer is assigned as follows:
\itemize{
\item \code{-1}: Job not found. This may be a false negative since Slurm may still
be scheduling the job.
\item \code{0}: Job completed. In this case all the components of the job
have finalized without any errors.
\item \code{1}: Some jobs are still running/pending.
\item \code{2}: One or more jobs have failed.
}
}
\description{
Check the State (status) of a Slurm JOB
}
\details{
Currently, the function state classifies jobs according to whether
these match the following state:
\itemize{
\item \code{done}    "CD",
\item \code{failed}  "BF", "CD", "DL", "F", "NF", "PR", "SE", "TO",
\item \code{running} "R",
\item \code{pending} "PD"
}
}
\seealso{
Other utilities: \code{\link{Slurm_clean}},
  \code{\link{Slurm_env}}, \code{\link{Slurm_log}},
  \code{\link{WhoAmI}}, \code{\link{parse_flags}},
  \code{\link{snames}}

Other post submission: \code{\link{Slurm_clean}},
  \code{\link{Slurm_collect}}, \code{\link{Slurm_log}}
}
\concept{post submission}
\concept{utilities}
\keyword{datasets}
