% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{status}
\alias{status}
\alias{status.slurm_job}
\alias{JOB_STATE_CODES}
\alias{status.default}
\title{Check the Status of a Slurm JOB}
\format{An object of class \code{list} of length 4.}
\usage{
status(x)

\method{status}{slurm_job}(x)

JOB_STATE_CODES

\method{status}{default}(x)
}
\arguments{
\item{x}{Either a Job id, or an object of class \code{slurm_job}.}
}
\value{
An integer with an attribute \code{sacct}
Codes:
\itemize{
\item \code{-1}: Job not found. This may be a false negative since Slurm may still
be scheduling the job.
\item \code{0}: Job completed. In this case all the components of the job
have finalized without any errors.
\item \code{1}: Some jobs are still running/pending.
\item \code{2}: One or more jobs have failed.
}
}
\description{
Check the Status of a Slurm JOB
}
\seealso{
Other Utility: \code{\link{parse_flags}},
  \code{\link{snames}}
}
\concept{Utility}
\keyword{datasets}
