% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{status}
\alias{status}
\alias{status.slurm_job}
\alias{status.default}
\title{Check the Status of a Slurm JOB}
\usage{
status(x)

\method{status}{slurm_job}(x)

\method{status}{default}(x)
}
\arguments{
\item{x}{Either a Job id, or an object of class \code{slurm_job}.}
}
\value{
An integer with an attribute \code{sacct}
Codes:
}
\description{
Check the Status of a Slurm JOB
}
