% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WhoAmI}
\alias{WhoAmI}
\alias{whoami}
\title{Information about where jobs are submitted}
\usage{
WhoAmI()

whoami()
}
\value{
A character vector with the corresponding system environment variables'
values.
}
\description{
This returns a named vector with the following variables:
\Sexpr[stage=build]{paste(names(slurmR::WhoAmI()), collapse = ", ")}
}
\details{
\code{whoami} is just an alias of \code{WhoAmI}.
}
\seealso{
Other utilities: 
\code{\link{Slurm_clean}()},
\code{\link{Slurm_env}()},
\code{\link{Slurm_log}()},
\code{\link{parse_flags}()},
\code{\link{snames}()},
\code{\link{status}()}
}
\concept{utilities}
