% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sbatch_list.R
\name{write_sbatch_list}
\alias{write_sbatch_list}
\title{Write a list of sbatch files to folder}
\usage{
write_sbatch_list(sbatch_list, path = "", overwrite = FALSE)
}
\arguments{
\item{sbatch_list}{A \code{list} with \code{sbatch} objects.}

\item{path}{Where to write slurm files.}

\item{overwrite}{Should files be overwritten?}
}
\description{
Write a list of sbatch files to folder
}
\details{
Function that write out sbatch files together with a \code{run_all.sh} 
bash script that start all batch scripts.
}

